Modularize validation functionality :tada:
- extracting rules & uniqueProp module outside of main module
